Generalize LayerNorm1P
